Remove blank line check in _get_symtab().
In the file processing loop in this function, there is was a check
for a blank line. It never succeeds because the lines have at least
a newline char at the end, and if it *did* succeed, there are two
'blank' lines right at the beginning of the xobjdump output, so this
would cause the rest of the file to be skipped. So I'm removing the
test and the 'break' statement. It also runs quicker, then :-)