Added cmake support & assertive comparison tester
Earlier work fell behind develop, so existing work has been reworked to match
the emerging style and direction of the repository.

In _build(...), specifying `cmake` now changes behaviour; a
key assumption is that CMakeLists.txt exists
in a superdirectory of the xe_path, and that directory also holds /bin/.

do_run_pyxsim now supports plugins of class XsiPlugin.

pyxsim.py is now guarded against XCC_EXEC_PREFIXes that do not end in /

AssertiveComparisonTester has been added;
this is used in essentially the same way as ComparisonTester, but
uses asserts rather than writing errors to stdout.
Should be more useful for pytest-based frameworks.
Personal opinion is that it's also more readable but that's by the by.